LPS-58999 Move Facebook SSO exceptions to API module
