+++ modules/apps/portal-security-sso/portal-security-sso-facebook-connect-api/build.gradle
@@ -1,4 +1,5 @@
 dependencies {
 	compileOnly group: "biz.aQute.bnd", name: "biz.aQute.bndlib", version: "3.1.0"
+	compileOnly group: "com.liferay.portal", name: "com.liferay.portal.kernel", version: "default"
 	compileOnly project(":apps:static:portal-configuration:portal-configuration-metatype-api")
 }
