+++ modules/apps/portal-security-sso/portal-security-sso-facebook-connect-api/src/main/java/com/liferay/portal/security/sso/facebook/connect/exception/StrangersNotAllowedException.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.login.authentication.facebook.connect.web.internal.portlet.action;
+package com.liferay.portal.security.sso.facebook.connect.exception;
 
 import com.liferay.portal.kernel.exception.PortalException;
