+++ modules/apps/portal-security-sso/portal-security-sso-facebook-connect-api/src/main/resources/com/liferay/portal/security/sso/facebook/connect/exception/packageinfo
@@ -1,0 +1,1 @@
+version 1.0.0
